requirements.txt
asgiref==3.2.10
DateTime==4.3
Django==3.1.2
django-webpack-loader==0.7.0
djangorestframework==3.12.1
more-itertools==8.2.0
pytz==2021.1
sqlparse==0.4.1
zope.interface==5.2.0
